requirements.txt
streamlit
python-dotenv
langgraph
langchain
langchain_core
langsmith
langchain-groq

google-search-results
duckduckgo-search
langchain-nvidia-ai-endpoints
faiss-cpu
langchain-community
langchain-text-splitters
PyPDF2
python-dotenv
pydantic


